refactor: move TooltipController to component-base (#4429)
