+++ packages/component-base/src/tooltip-controller.d.ts
@@ -3,7 +3,7 @@
  * Copyright (c) 2022 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { SlotController } from '@vaadin/component-base/src/slot-controller.js';
+import { SlotController } from './slot-controller.js';
 
 /**
  * A controller that manages the slotted tooltip element.
